split web module
